backend/requirements.txt
uvicorn==0.23.2
pydantic==1.10.12
numpy==1.24.4
bcrypt==4.0.1
fastapi-mail==1.0.8
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
sqlalchemy==2.0.19
soundfile==0.12.1
pyjwt==2.8.0
passlib==1.7.4
librosa==0.10.0
fastapi==0.75.2
python-dotenv==1.0.0
pydub==0.25.1
matplotlib==3.7.2
PyWavelets==1.4.1
pandas==2.0.3
transformers==4.31.0